Moved render placements to a different (earlier) location
